AI App Development/explore-openai/prompt-engineering/pyproject.toml
[tool.poetry]
name = "explore-open-ai"
version = "0.1.0"
description = "Project for exploring open ai."
authors = ["Dilip.Sundarraj <dilip.sundarraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.70.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
yfinance = "^0.2.43"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
